backend/app/utils/file_extractors/docx_ext.py
"""DOCX text extraction using python-docx."""
from docx import Document
from io import BytesIO


def extract_text_from_docx(file_bytes: bytes) -> str:
    """Extract text content from a DOCX file."""
    doc = Document(BytesIO(file_bytes))
    paragraphs = [para.text for para in doc.paragraphs if para.text.strip()]
    return "\n\n".join(paragraphs)
